requirements.txt
# Core dependencies for Hybrid RAG System
# Python >= 3.8

# NLP and Transformers
transformers>=4.30.0
sentence-transformers>=2.2.0
nltk>=3.8.0
scikit-learn>=1.3.0

# Vector search and indexing
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
rank-bm25>=0.2.2

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Web framework and UI
streamlit>=1.28.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Data and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
jsonlines>=4.0.0

# Development and testing
pytest>=7.4.0
